Added some premade buttons and textviews
Added a bunch of buttons. Id of buttons is as follows: "button1" and
then upp to "button8".
Added a bunch of textview's to go along with the buttons. Id of them are
as follows "textview1" to "textview6".
NOTE! Change these buttons and texviews names to something better when
they have an actual function.